feat: add CLI interface for design tokens
