+++ tokens/package.json
@@ -4,7 +4,7 @@
   "description": "Paragon's design system based on Style Dictionary approach.",
   "main": "build-tokens.js",
   "scripts": {
-    "build-tokens": "node build-tokens.js",
+    "build-tokens": "node build-tokens.js --build-dir ../scss/core/css/",
     "replace-variables-usage": "node replace-variables.js -p ../src/ -t usage",
     "replace-variables-definition": "node replace-variables.js -p ../src -t definition",
     "build-scss-to-css-map": "node map-scss-to-css.js"
@@ -13,7 +13,7 @@
   "license": "ISC",
   "dependencies": {
     "chroma-js": "^2.4.2",
-    "commander": "^7.1.0",
+    "commander": "^7.2.0",
     "style-dictionary": "^3.7.1"
   }
 }
